pyproject.toml
[tool.poetry]
name = "questbackend"
version = "0.1.0"
description = ""
authors = ["shubham zade <shubham.zade@dentsu.com>"]
readme = "README.md"
package-mode = false   # prevent poetry from trying to package your app

[tool.poetry.dependencies]
python = ">=3.13"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = { extras = ["standard"], version = ">=0.35.0,<0.36.0" }
sqlalchemy = ">=2.0.43,<3.0.0"
psycopg = { extras = ["binary"], version = ">=3.2.9,<4.0.0" }
python-dotenv = ">=1.1.1,<2.0.0"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
snowflake-connector-python = "^3.12.2"
snowflake-sqlalchemy = "^1.5.2"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
